% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbChrom.R
\docType{data}
\name{carbChrom}
\alias{carbChrom}
\title{Chromatogram dataset example}
\format{
An object of class \code{data.frame} with 4119 rows and 2 columns.
}
\usage{
data(carbChrom)
}
\description{
Data from Carbocisteine UV Chromatogram
}
\examples{
data(carbChrom)
}
\keyword{datasets}
