% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_MAP_estimate.R
\name{create_MAP_estimate}
\alias{create_MAP_estimate}
\title{Create MAP Estimates for an individual using an mrgsolve model}
\usage{
create_MAP_estimate(
  x,
  tdm_data,
  ycol_name = "CONC",
  dvcol_name = "IPRED",
  init_etas
)
}
\arguments{
\item{x}{mrgsim model to be used in the simulation (must contain parameters called ETA1 ... ETAX)}

\item{tdm_data}{contains covariates, dosing events and observed concentration}

\item{ycol_name}{column name of the observed concentration in tdm_data}

\item{dvcol_name}{name of simulation result}

\item{init_etas}{named vector of initial values for MAP bayes estimation (number and name of ETAS must conform with the model, names: ETA1 ... ETAX)}
}
\description{
Create MAP Estimates for an individual using an mrgsolve model
}
