% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pred_int.R
\name{create_pred_int}
\alias{create_pred_int}
\title{Create predictions intervals from mrgsim result}
\usage{
create_pred_int(
  x,
  dv_column = "IPRED",
  time_column = "time",
  id_column = "ID",
  evid_column = "evid"
)
}
\arguments{
\item{x}{mrgsim result (must contain an event-ID column)}

\item{dv_column}{name of the column containing the dependent variable}

\item{id_column}{name of the column containing ID}

\item{evid_column}{name of the column containing event-ID}

\item{time}{name of the column containing the longitudinal (independent) variable}
}
\value{
a data.frame containing the independent variable, the 2.5th, 5th, 25th, 50th, 75th, 95th and 97.5th percentile of the dependent variable
}
\description{
Create predictions intervals from mrgsim result
}
\examples{
data(ruxSim)
create_pred_int(ruxSim)
}
