% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pred_plot.R
\name{create_pred_plot}
\alias{create_pred_plot}
\title{Create a plot with prediction intervals from mrgsim result and optional observed data}
\usage{
create_pred_plot(
  x,
  intervall_colour = "red",
  median_colour = "darkblue",
  obs_colour = "black",
  obs_shape = 1,
  obs_size = 3,
  y_lab = "Plasma Concentration [ng/mL]",
  x_lab = "Time [h]",
  plot_title = "PK-Curve",
  plot_subtitle = "oral",
  text_size = 14,
  obs = NULL
)
}
\arguments{
\item{x}{mrgsim result (must contain an event-ID column)}

\item{median_colour}{specify colour of the median line}

\item{obs_colour}{specify colour of the observations}

\item{obs_shape}{specify shape of the observations}

\item{y_lab}{specify label of the y-axis}

\item{x_lab}{specify label of x-axis}

\item{plot_title}{specify plot title}

\item{plot_subtitle}{specify plot subtitle}

\item{text_size}{specify text size for labels and title}

\item{obs}{a data.frame containing two columns called \code{x} and \code{y} with independent and dependent variable of the observations}
}
\value{
The result is a ggplot object
}
\description{
Create a plot with prediction intervals from mrgsim result and optional observed data
}
