% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenChromatogram.R
\docType{class}
\name{OpenChromatogram}
\alias{OpenChromatogram}
\title{OpenChromatogram class}
\value{
an R6 class object of OpenChromatogram

a ggplot2 object

a data.frame with three columns, first and second derivative and peak detection
}
\description{
OpenChromatogram class

OpenChromatogram class
}
\details{
This an open source format for dealing with chromatograms in R
}
\examples{
data(carbChrom)
OpenChromatogram$new("UV210nm", carbChrom$mAU, carbChrom$Time,
                    "min", "mAU") -> chrom
plot(chrom)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detector}}{String indicating the detector type}

\item{\code{time_trace}}{String indicating the name of the time domain trace}

\item{\code{signal_trace}}{String indicating the name of signal trace}

\item{\code{unit_time}}{String indicating the time unit}

\item{\code{unit_signal}}{String indicating the detector signal unit}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OpenChromatogram$new()}}
\item \href{#method-plot}{\code{OpenChromatogram$plot()}}
\item \href{#method-peak_detect}{\code{OpenChromatogram$peak_detect()}}
\item \href{#method-clone}{\code{OpenChromatogram$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
OpenChromatogram is used for plotting chromatogram data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenChromatogram$new(
  detector = NULL,
  signal_trace = NULL,
  time_trace = NULL,
  unit_time,
  unit_signal
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{detector}}{String object  indicating the detector type}

\item{\code{signal_trace}}{String object indicating the name of the time domain trace}

\item{\code{time_trace}}{String object indicating the name of signal trace}

\item{\code{unit_time}}{String object indicating the time unit}

\item{\code{unit_signal}}{String object indicating the detector signal unit}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Creates a ggplot2 object from the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenChromatogram$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-peak_detect"></a>}}
\if{latex}{\out{\hypertarget{method-peak_detect}{}}}
\subsection{Method \code{peak_detect()}}{
Integrates the chromatogram using a given threshold
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenChromatogram$peak_detect(threshold = 0.001)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{threshold for peak detection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenChromatogram$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
