% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruxSim.R
\docType{data}
\name{ruxSim}
\alias{ruxSim}
\title{mrgsim dataset example}
\format{
An object of class \code{mrgsims} with 36200 rows and 18 columns.
}
\usage{
data(ruxSim)
}
\description{
Data from Ruxolitinib population PK Simulation
}
\examples{
data(ruxSim)
}
\keyword{datasets}
